Greatly improve code quality, fix a bug during movie import.
